refactor(site): siteディレクトリをbrandsとdecorationsに再編成
- Avatar, Profile, FullProfile, SnsList を brands/profiles/ に移動
- SpeechBubble を decorations/speech-bubbles/ に移動
- テストファイルとスクリーンショットも対応するディレクトリに移動
- 各機能グループに index.ts を作成
- site/ ディレクトリを削除
- 6ファイルのインポートパスを新しい構造に更新
- docs/components-reference.md を更新
