+++ docs/components-reference.md
@@ -3,7 +3,7 @@
 このドキュメントは `src/components` 以下に存在する共通 React コンポーネントの一覧です。
 新しいコンポーネントを作成する前に、既存の実装がないか確認してください。
 
-## ディレクトリ構造
+## Components ディレクトリ構造
 
 コンポーネントは**カテゴリ/機能グループ**の 2 階層で配置します。
 機能グループは単一のコンポーネントではなく、関連するコンポーネントをまとめた意味のある単位です。
@@ -11,12 +11,14 @@
 ```text
 src/components/
 ├── brands/
-│   └── opengraph-image/       # OGP画像生成
+│   ├── opengraph-image/       # OGP画像生成
+│   └── profiles/              # サイトプロフィール
 ├── controls/
 │   └── pagination/            # ページネーション制御
 ├── decorations/
 │   ├── effects/               # モーション・視覚効果
-│   └── scroll-shadow/         # スクロールシャドウ
+│   ├── scroll-shadow/         # スクロールシャドウ
+│   └── speech-bubbles/        # 吹き出し
 ├── displays/                  # 将来用（データ表示コンポーネント）
 ├── feedback/
 │   └── message/               # メッセージ表示
@@ -34,12 +36,6 @@ src/components/
 ├── navigations/
 │   └── link/                  # リンク関連コンポーネント
 ├── prefabs/                   # 将来用（特定用途向けパターン）
-├── site/                      # サイト固有コンポーネント（移動予定）
-│   ├── avatar.tsx
-│   ├── full-profile.tsx
-│   ├── profile.tsx
-│   ├── sns-list.tsx
-│   └── speech-bubble.tsx
 ├── typography/
 │   # 以下は直下に配置（サブディレクトリ化予定）
 │   ├── heading.tsx            # 見出し
@@ -87,8 +83,7 @@ import { Pagination } from '@/components/controls'
 | `layouts`     | ナビゲーションバー、コンテナ、フッターなど、ページの骨格・構造を定義するコンポーネント |
 | `misc`        | Google Analytics、テーマ切替など、特定カテゴリに属さないサイト機能                     |
 | `navigations` | リンク、ナビゲーションなど、ページ間・ページ内の移動を担う UI                          |
-| `prefabs`     | ブックマークリスト、タグ、サイトプロフィールなど、特定用途向けに構成した再利用パターン |
-| `site`        | サイト固有のプロフィール関連コンポーネント（将来的に `prefabs` へ移動予定）            |
+| `prefabs`     | ブックマークリスト、タグなど、特定用途向けに構成した再利用パターン                     |
 | `typography`  | 見出し、段落など、テキスト表示のためのコンポーネント                                   |
 | `ui`          | shadcn/ui ベースの基本 UI パーツ。他のコンポーネントの構成要素として使用               |
 
@@ -96,9 +91,10 @@ import { Pagination } from '@/components/controls'
 
 ## brands
 
-| ディレクトリ             | 説明                                                                                                                       |
-| ------------------------ | -------------------------------------------------------------------------------------------------------------------------- |
-| `brands/opengraph-image` | OGP 用動的画像生成。Next.js ImageResponse でソーシャルメディア共有用画像を生成。主なエクスポート: `DocumentOpengraphImage` |
+| ディレクトリ             | 説明                                                                                                                                           |
+| ------------------------ | ---------------------------------------------------------------------------------------------------------------------------------------------- |
+| `brands/opengraph-image` | OGP 用動的画像生成。Next.js ImageResponse でソーシャルメディア共有用画像を生成。主なエクスポート: `DocumentOpengraphImage`                     |
+| `brands/profiles`        | サイトプロフィール関連。`Avatar`（アバター）、`Profile`（簡易プロフィール）、`FullProfile`（詳細プロフィール）、`SnsList`（SNSリンク）を提供。 |
 
 ## controls
 
@@ -108,10 +104,11 @@ import { Pagination } from '@/components/controls'
 
 ## decorations
 
-| ディレクトリ                | 説明                                                                                                              |
-| --------------------------- | ----------------------------------------------------------------------------------------------------------------- |
-| `decorations/effects`       | モーション・視覚効果コンポーネント。`FlowingGlow`（流れるグロー効果）、`HoverShimmer`（ホバーシマー効果）を提供。 |
-| `decorations/scroll-shadow` | スクロール可能領域にグラデーション影を表示しスクロール方向を視覚化。主なエクスポート: `ScrollShadow`              |
+| ディレクトリ                 | 説明                                                                                                              |
+| ---------------------------- | ----------------------------------------------------------------------------------------------------------------- |
+| `decorations/effects`        | モーション・視覚効果コンポーネント。`FlowingGlow`（流れるグロー効果）、`HoverShimmer`（ホバーシマー効果）を提供。 |
+| `decorations/scroll-shadow`  | スクロール可能領域にグラデーション影を表示しスクロール方向を視覚化。主なエクスポート: `ScrollShadow`              |
+| `decorations/speech-bubbles` | 吹き出し風のコンテナ。`SpeechBubble` を提供。キャラクターが喋っているような吹き出しを描画。                       |
 
 ## displays
 
@@ -160,12 +157,6 @@ import { Pagination } from '@/components/controls'
 | ------------ | ------------------------------------------------------------------------------------------------------------------ |
 | `prefabs/`   | 将来用のカテゴリ。特定用途向けに構成した再利用パターン（複数コンポーネントの組み合わせ）を配置予定。現在は空です。 |
 
-## site
-
-| ディレクトリ | 説明                                                                                                                                                 |
-| ------------ | ---------------------------------------------------------------------------------------------------------------------------------------------------- |
-| `site/`      | サイト固有のプロフィール関連コンポーネント。`Avatar`、`FullProfile`、`Profile`、`SnsList`、`SpeechBubble` を提供。将来的に適切なカテゴリへ移動予定。 |
-
 ## typography
 
 | ディレクトリ  | 説明                                                                                          |
@@ -180,12 +171,12 @@ import { Pagination } from '@/components/controls'
 
 ---
 
-# Features リファレンス
+## Features リファレンス
 
 `src/features` 以下には、このプロジェクト専用の機能に関するコンポーネントを配置します。
 汎用的な UI コンポーネントではなく、特定のビジネスロジックやドメイン知識に依存するコンポーネント群です。
 
-## ディレクトリ構造
+### Features ディレクトリ構造
 
 ```text
 src/features/
@@ -214,7 +205,7 @@ src/features/
     └── index.ts
 ```
 
-## 機能一覧
+### Features 機能一覧
 
 | ディレクトリ | 説明                                                                                                                                                                                                                                                                                                                                    |
 | ------------ | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
