+++ src/components/brands/profiles/full-profile.tsx
@@ -1,7 +1,10 @@
+import { SpeechBubble } from '@/components/decorations/speech-bubbles'
 import { Link } from '@/components/navigations/link'
-import { Avatar, SnsList, SpeechBubble } from '@/components/site'
 import { Heading, Paragraph as P } from '@/components/typography'
 import { SITE_DESC } from '@/constants'
+
+import Avatar from './avatar'
+import SnsList from './sns-list'
 
 
 /**
