+++ src/components/brands/profiles/index.ts
@@ -1,0 +1,4 @@
+export { default as Avatar } from './avatar'
+export { default as FullProfile } from './full-profile'
+export { default as Profile } from './profile'
+export { default as SnsList } from './sns-list'
