+++ src/components/brands/profiles/sns-list.tsx
