+++ src/components/decorations/speech-bubbles/index.ts
@@ -1,0 +1,1 @@
+export { default as SpeechBubble } from './speech-bubble'
