+++ src/components/decorations/speech-bubbles/speech-bubble.tsx
