--- src/components/site/index.ts
@@ -1,5 +1,0 @@
-export { default as SnsList } from './sns-list'
-export { default as Avatar } from './avatar'
-export { default as Profile } from './profile'
-export { default as FullProfile } from './full-profile'
-export { default as SpeechBubble } from './speech-bubble'
